Update managed-application-router-589a239.md
As the SAP Launchpad Service is not existing anymore I updated with the new SAP Build Work Zone, standard edition service to avoid confusion. This is also matching with the links provided later in the text. There I provided the direct link to the WZ Std help page and not the redirect from the launchpad article.